Refactor LegoContainer
